+++ src/main/java/com/titaniumtigers4829/TigerHelpers.java
@@ -3,6 +3,7 @@
 package com.titaniumtigers4829;
 
 import com.titaniumtigers4829.data.IMUData;
+import com.titaniumtigers4829.data.IMUData.IMUMode;
 import com.titaniumtigers4829.data.PoseEstimate;
 import com.titaniumtigers4829.data.PoseEstimate.Botpose;
 import com.titaniumtigers4829.data.RawFiducial;
@@ -390,19 +391,6 @@ public class TigerHelpers {
   }
 
   /**
-   * Checks if a PoseEstimate is valid. If this is true, it means the PoseEstimate is has valid data
-   * from at least one april tag.
-   *
-   * @param poseEstimate The PoseEstimate to check
-   * @return True if the PoseEstimate is valid, false otherwise
-   */
-  public static boolean validPoseEstimate(PoseEstimate poseEstimate) {
-    return poseEstimate != null
-        && poseEstimate.rawFiducials != null
-        && poseEstimate.rawFiducials.length != 0;
-  }
-
-  /**
    * Sets the network table entry for the blue-side, MegaTag1 botpose data. This is useful for
    * setting values for unit testing. The {@link PoseEstimate} does not contain values for the z
    * coordinate, roll, and pitch, so these will be set to 0.
@@ -554,70 +542,6 @@ public class TigerHelpers {
    */
   public static void setRobotOrientation(String limelightName, double yaw) {
     setRobotOrientation(limelightName, yaw, 0, 0, 0, 0, 0);
-  }
-
-  /**
-   * Enum representing different IMU usage modes for robot orientation or localization. Defines how
-   * internal and external IMUs are utilized, including seeding and convergence assistance.
-   */
-  public enum IMUMode {
-    /**
-     * Use only an external IMU orientation set by setRobotOrientation() for orientation data. No
-     * interaction with the internal IMU.
-     */
-    EXTERNAL_IMU(0),
-
-    /**
-     * Use an external IMU orientation set by setRobotOrientation() as the primary source and seed
-     * the internal IMU with its data. This should be used for "zeroing" the internal IMU. The
-     * internal IMU is initialized or calibrated using external IMU values.
-     */
-    EXTERNAL_IMU_SEED_INTERNAL(1),
-
-    /** Use only the internal IMU for orientation data. No reliance on an external IMU. */
-    INTERNAL_IMU(2),
-
-    /**
-     * Use the internal IMU with MT1-assisted convergence. MegaTag1 provides additional data to
-     * improve internal IMU accuracy or stability.
-     */
-    INTERNAL_MT1_ASSISTED(3),
-
-    /**
-     * Use the internal IMU with external IMU-assisted convergence. The external IMU provides
-     * supplementary data to enhance internal IMU performance.
-     */
-    INTERNAL_EXTERNAL_ASSISTED(4);
-
-    private final int modeValue;
-
-    IMUMode(int modeValue) {
-      this.modeValue = modeValue;
-    }
-
-    /**
-     * Returns the integer value associated with this IMU mode.
-     *
-     * @return The mode value (0 for EXTERNAL_IMU, 1 for EXTERNAL_IMU_SEED_INTERNAL, etc.)
-     */
-    public int getModeValue() {
-      return modeValue;
-    }
-
-    /**
-     * Returns the IMUMode corresponding to the given integer value.
-     *
-     * @param value The integer value to convert to an IMUMode
-     * @return The matching IMUMode, or null if no match is found
-     */
-    public static IMUMode fromValue(int value) {
-      for (IMUMode mode : IMUMode.values()) {
-        if (mode.modeValue == value) {
-          return mode;
-        }
-      }
-      return null;
-    }
   }
 
   /**
